# gen/openapiv2/protos/my_super_service.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "protos/my_super_service.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "MySuperService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/your.service.v1.MySuperService/AddNotam": {
      "post": {
        "operationId": "MySuperService_AddNotam",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1AddNotamOutput"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1AddNotamInput"
            }
          }
        ],
        "tags": [
          "MySuperService"
        ]
      }
    },
    "/your.service.v1.MySuperService/ListNotams": {
      "post": {
        "operationId": "MySuperService_ListNotams",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListNotamsOutput"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1ListNotamsInput"
            }
          }
        ],
        "tags": [
          "MySuperService"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1AddNotamInput": {
      "type": "object",
      "properties": {
        "notam": {
          "$ref": "#/definitions/v1Notam"
        }
      }
    },
    "v1AddNotamOutput": {
      "type": "object",
      "properties": {
        "notam": {
          "$ref": "#/definitions/v1Notam"
        }
      }
    },
    "v1ListNotamsInput": {
      "type": "object",
      "properties": {
        "limit": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "v1ListNotamsOutput": {
      "type": "object",
      "properties": {
        "notams": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1Notam"
          }
        }
      }
    },
    "v1Notam": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    }
  }
}